Add files via upload
We see that all the employee’s details are returned in an HTML tabular format. Hence, we were 
able to perform the same attack as SQL Injection attack from web page. The CLI commands can help in automating the 
attack, where Web UI don’t. One major change from the web UI was to encode the special 
characters in the HTTP request in the curl command. We use the following: Space - %20; Hash 
(#) - %23 and Single Quote (‘) - %27.